+++ generator/api/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/api/simplification/SimplificationTreeTest.kt
@@ -70,15 +70,5 @@ class SimplificationTreeTest {
             )
         )
     }
-
-    private fun <T> assertTreeEquals(expected: SimplificationTree<T>, actual: SimplificationTree<T>) {
-        assertEquals(expected.root, actual.root)
-        val actualBranches = actual.branches.toList()
-        val expectedBranches = expected.branches.toList()
-
-        assertEquals(expectedBranches.size, actualBranches.size)
-        expectedBranches.forEachIndexed { index, expectedTree ->
-            assertTreeEquals(expectedTree, actualBranches[index])
-        }
-    }
 }
+
